+++ src/index.js
@@ -5,7 +5,7 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from './store.js'
+import store from './store/store.js'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
