+++ src/routes/Login.css
@@ -1,39 +1,80 @@
 .loginForm {
+  max-width: 500px; /* 박스 크기 조정 */
+  margin: 15px auto;
+  padding: 2.5rem; /* 패딩 조정 */
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  background-color: #ffffff;
+  position: relative; /* Relative positioning for absolute positioning of link */
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1); /* 그림자 추가 */
+}
+
+.loginTitle {
+  font-size: 2.5rem; /* 제목 글씨 크기 조정 */
+  color: #00796b; /* 제목 색상 */
+  text-align: center; /* 제목 중앙 정렬 */
+  margin-bottom: 2rem; /* 제목과 폼 간의 여백 */
+}
+
+.loginForm .form-control {
+  font-size: 1.2rem; /* 입력 필드 글씨 크기 조정 */
+  border: none;
+  border-bottom: 1px solid #e0e0e0;
+  padding: 0.75rem; /* 패딩 조정 */
+  outline: none;
   width: 100%;
-  height: 70%;
-  max-width: 550px; /* 최대 너비 설정 */
-  padding: 20px; /* 폼 내부 여백 */
-  border: 2px solid #ced4da; /* 테두리 스타일 */
-  border-radius: 5px; /* 둥근 테두리 */
-  background-color: #ffffff; /* 배경 색상 */
-  box-shadow: 0 2px 4px rgba(141, 126, 126, 0.1); /* 그림자 효과 */
-  position: absolute; /* 절대 위치 지정 */
-  top: 50%; /* 세로 중앙 정렬 */
-  left: 50%; /* 가로 중앙 정렬 */
-  transform: translate(-50%, -50%); /* 중앙 정렬 조정 */
+  margin-bottom: 1.5rem; /* 입력 필드 간의 여백 */
+}
+
+.loginForm .form-control:focus {
+  border-bottom: 1px solid #b0b0b0;
+  color: #00796b;
 }
 
 .btn-primary {
-  width: 100%;       /* form의 가로 길이에 맞춤 */
-  padding: 4px;     /* 버튼의 높이를 늘리기 위해 패딩 조절 */
-  font-size: 20px;  /* 텍스트 크기 조절 */
+  margin-top: 1.5rem; /* 버튼과 입력 필드 간의 여백 */
+  width: 100%;
+  background-color: #00796b;
+  border: none;
+  color: #ffffff;
+  padding: 0.75rem; /* 버튼 패딩 조정 */
+  font-size: 1.2rem; /* 버튼 글씨 크기 조정 */
+  border-radius: 4px;
+  cursor: pointer;
 }
 
+.btn-primary:hover {
+  background-color: #004d40;
+}
 
-/* Login.css */
+.kakaobtn {
+  text-align: center;
+  margin-top: 1.5rem; /* 카카오 버튼과 로그인 폼 간의 여백 */
+}
+
+.kakaobtn img {
+  max-width: 100%;
+  border: none;
+  cursor: pointer;
+}
+
 .findRemove {
-  list-style: none; /* 기본 리스트 스타일 제거 */
+  list-style-type: none;
   padding: 0;
-  margin: 0 auto;
-  display: inline-flex; /* 플렉스박스 사용 */
-  gap: 10px; /* 아이템 간 간격 조절 */
-  font-size: 12px;
-  margin-top: 10px;
-  margin-left: 80px;
+  margin-top: 1.5rem; /* 회원가입 링크와 카카오 버튼 간의 여백 */
 }
 
-.findRemove li:not(:last-child)::after {
-  content: '|'; /* 구분자 추가 */
-  margin-left: 10px;
+.findRemove li {
+  position: absolute; /* Absolute positioning for precise placement */
+  bottom: 10px; /* 위치 조정 */
+  right: 10px; /* 위치 조정 */
+  text-align: right;
+  font-size: 1rem; /* 회원가입 링크 글씨 크기 조정 */
+  color: #00796b;
+  cursor: pointer;
+  text-decoration: underline;
 }
 
+.findRemove li:hover {
+  color: #004d40;
+}
