+++ src/routes/Login.js
@@ -1,140 +1,88 @@
-import './Login.css';
-import { useState } from 'react';
+import React, { useState } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
+import './Login.css'; // CSS 파일을 임포트
 import kakaoImg from '../images/kakao_login.png';
 import { KAKAO_AUTH_URL } from '../oauth/Oauth';
 
 function Login() {
-  const code = new URL(window.location.href).searchParams.get('code'); // 인가코드 추출
-  const navigate = useNavigate();
+    const navigate = useNavigate();
 
-  let [userId, setUserId] = useState('');
-  let [userPw, setUserPw] = useState('');
+    const [userId, setUserId] = useState('');
+    const [userPw, setUserPw] = useState('');
 
-  // 아이디와 비밀번호 입력받는 값 저장
-  const userIdChange = (e) => setUserId(e.target.value);
-  const userPwChange = (e) => setUserPw(e.target.value);
+    const userIdChange = (e) => setUserId(e.target.value);
+    const userPwChange = (e) => setUserPw(e.target.value);
 
-  //formData를 JSON으로 변환하는 함수
-  const formDataToJSON = (formData) => {
-    const obj = {};
-    formData.forEach((value, key) => {
-      obj[key] = value;
-    });
-    return obj;
-  };
+    const onSubmit = (e) => {
+        e.preventDefault();
 
-  // 로그인 버튼 누르면 로그인 기능 실행
-  const onSubmit = (e) => {
-    e.preventDefault();
+        axios
+            .post('http://localhost:8080/api/members/login', {
+                userId: userId,
+                userPw: userPw,
+            })
+            .then((res) => {
+                sessionStorage.setItem('user', JSON.stringify(res.data));
+                sessionStorage.setItem('isLoggedIn', true);
+                navigate('/');
+            })
+            .catch(() => {
+                alert('로그인에 실패하였습니다.');
+                window.location.reload();
+            });
+    };
 
-    let formData = new FormData();
-    formData.append('userId', userId);
-    formData.append('userPw', userPw);
-
-    const data = formDataToJSON(formData);
-
-    //포스트 요청으로 아이디 비밀번호 값 보내기
-    axios
-      .post('http://localhost:8080/api/members/login', {
-        userId: userId,
-        userPw: userPw,
-      })
-      .then((res) => {
-        sessionStorage.setItem('user', JSON.stringify(res.data)); //session에 값 추가
-        sessionStorage.setItem('isLoggedIn', true);
-        navigate('/');
-      })
-      .catch(function (error) {
-        alert('로그인에 실패하였습니다.');
-        window.location.reload();
-      });
-  };
-
-  const handleKakaoLogin = () => {
-    // 카카오 로그인 관련 로직 처리
-  };
-  return (
-    <>
-      <div>
-        <form
-          className="loginForm"
-          onSubmit={onSubmit}
-          encType="multipart/form"
-        >
-          {/* 아이디 입력 칸 */}
-          <div className="mb-3">
-            <input
-              type="text"
-              value={userId}
-              onChange={userIdChange}
-              className="form-control"
-              placeholder="아이디"
-            />
-          </div>
-
-          {/* 비밀번호 입력칸 */}
-          <div className="mb-3">
-            <input
-              type="text"
-              value={userPw}
-              onChange={userPwChange}
-              className="form-control"
-              placeholder="비밀번호"
-            />
-          </div>
-
-          <div className="mb-3 form-check">
-            {/* 아이디 저장 체크박스 */}
-            <input type="checkbox" className="member" id="checkId" />
-            <label htmlFor="checkId">아이디 저장</label>
-
-            {/* 자동 로그인 체크박스 */}
-            <input type="checkbox" className="member" id="autoLogin" />
-            <label htmlFor="autoLogin">자동로그인</label>
-          </div>
-
-          {/* 로그인 버튼 */}
-          <button type="submit" className="btn btn-primary" value="user login">
-            로그인
-          </button>
-
-          {/* 찾기,휴면해제,회원가입 */}
-          <ul className="findRemove">
-            <li>아이디/비밀번호 찾기</li>
-            <li>휴면해제</li>
-            <li
-              onClick={() => {
-                navigate('/signUp');
-              }}
-              style={{ cursor: 'pointer' }}
+    return (
+        <div>
+            <form
+                className="loginForm"
+                onSubmit={onSubmit}
             >
-              회원가입
-            </li>
-          </ul>
+                {/* 로그인 제목 */}
+                <h2 className="loginTitle">로그인</h2>
 
-          <div style={{ textAlign: 'center', marginTop: '20px' }}>
-            <a
-              href={KAKAO_AUTH_URL}
-              className="kakaobtn"
-              style={{ display: 'inline-block' }}
-            >
-              <img
-                src={kakaoImg}
-                alt="카카오 로그인"
-                style={{
-                  padding: '0',
-                  border: 'none',
-                  cursor: 'pointer',
-                }}
-              />
-            </a>
-          </div>
-        </form>
-      </div>
-    </>
-  );
+                {/* 아이디 입력 칸 */}
+                <div className="mb-3">
+                    <input
+                        type="text"
+                        value={userId}
+                        onChange={userIdChange}
+                        className="form-control"
+                        placeholder="아이디"
+                        required
+                    />
+                </div>
+
+                {/* 비밀번호 입력칸 */}
+                <div className="mb-3">
+                    <input
+                        type="password"
+                        value={userPw}
+                        onChange={userPwChange}
+                        className="form-control"
+                        placeholder="비밀번호"
+                        required
+                    />
+                </div>
+
+                {/* 로그인 버튼 */}
+                <button type="submit" className="btn btn-primary">
+                    로그인
+                </button>
+
+                <div className="kakaobtn">
+                    <a href={KAKAO_AUTH_URL}>
+                        <img src={kakaoImg} alt="카카오 로그인" />
+                    </a>
+                </div>
+
+                <ul className="findRemove">
+                    <li onClick={() => navigate('/signUp')}>회원가입</li>
+                </ul>
+            </form>
+        </div>
+    );
 }
 
 export default Login;
